Prettification Pass. Made level look nice + Some fixes to buttons
I ran through the entire game and made everything look as nice as I
could. I also took some time to fix some issues I was having with the
buttons, and to improve the distortion behavior in the level. Puzzles
may not be working as intended, but mechanically, the game is solid.